requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.1.1
python-dotenv==1.0.0
werkzeug==2.3.7
sqlalchemy==2.0.21
pymysql==1.1.0
wtforms==3.0.1
email-validator==2.0.0
python-dateutil==2.8.2
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
httpx==0.25.0
pydantic==2.4.2
langchain==0.0.300
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
